refactor: move classes in handler and factory files to basemapper.py
