+++ osm_fieldwork/basemapper.py
@@ -21,38 +21,144 @@
 
 import argparse
 import concurrent.futures
-from io import BytesIO
 import logging
 import queue
 import re
 import sys
 import threading
+from io import BytesIO
 from pathlib import Path
-from typing import Union
+from typing import Tuple, Union
 
 import geojson
 import mercantile
 from cpuinfo import get_cpu_info
-from pmtiles.tile import (
-    Compression as PMTileCompression,
-)
-from pmtiles.tile import (
-    TileType as PMTileType,
-)
-from pmtiles.tile import (
-    zxy_to_tileid,
-)
+from pmtiles.tile import Compression as PMTileCompression
+from pmtiles.tile import TileType as PMTileType
+from pmtiles.tile import zxy_to_tileid
 from pmtiles.writer import Writer as PMTileWriter
 from pySmartDL import SmartDL
 from shapely.geometry import shape
 from shapely.ops import unary_union
-from osm_fieldwork.boundary_handler_factory import BoundaryHandlerFactory
+
 from osm_fieldwork.sqlite import DataFile, MapTile
 from osm_fieldwork.xlsforms import xlsforms_path
 from osm_fieldwork.yamlfile import YamlFile
 
 # Instantiate logger
 log = logging.getLogger(__name__)
+
+BoundingBox = Tuple[float, float, float, float]
+
+
+class BoundaryHandlerFactory:
+    """Factory class for creating boundary handlers based on the type of input boundary provided."""
+
+    def __init__(self, boundary: Union[str, BytesIO]):
+        """Initialize the BoundaryHandlerFactory with a boundary input.
+
+        Args:
+            boundary (Union[str, BytesIO]): The boundary input, either as a GeoJSON BytesIO object or a BBOX string.
+        """
+        if isinstance(boundary, BytesIO):
+            self.handler = BytesIOBoundaryHandler(boundary)
+        elif isinstance(boundary, str):
+            self.handler = StringBoundaryHandler(boundary)
+        else:
+            raise ValueError("Unsupported type for boundary parameter.")
+
+        self.boundary_box = self.handler.make_bbox()
+
+    def get_bounding_box(self) -> BoundingBox:
+        """Get bounding box.
+
+        Returns:
+            BoundingBox: The bounding box as a tuple (min_x, min_y, max_x, max_y).
+        """
+        return self.boundary_box
+
+
+class BoundaryHandler:
+    """A class to extract Bounding Box (BBOX) from various boundary representations."""
+
+    def make_bbox(self) -> BoundingBox:
+        """Extract and return the bounding box from the boundary representation.
+
+        Returns:
+        BoundingBox: The bounding box as a tuple (min_x, min_y, max_x, max_y).
+        """
+        pass
+
+
+class BytesIOBoundaryHandler(BoundaryHandler):
+    """Extracts BBOX from GeoJSON data stored in a BytesIO object."""
+
+    def __init__(self, boundary: BytesIO):
+        """Initialize the BytesIOBoundaryHandler with a BytesIO input."""
+        self.boundary = boundary
+
+    def make_bbox(self) -> BoundingBox:
+        """Extract and return the bounding box from the GeoJSON data.
+
+        Returns:
+            BoundingBox: The bounding box as a tuple (min_x, min_y, max_x, max_y).
+        """
+        log.debug(f"Reading geojson BytesIO : {self.boundary}")
+        # Rewind the BytesIO object to the beginning before passing it to geojson.load()
+        self.boundary.seek(0)
+        with self.boundary as buffer:
+            poly = geojson.load(buffer)
+
+        if "features" in poly:
+            geometry = shape(poly["features"][0]["geometry"])
+        elif "geometry" in poly:
+            geometry = shape(poly["geometry"])
+        else:
+            geometry = shape(poly)
+
+        if isinstance(geometry, list):
+            # Multiple geometries
+            log.debug("Creating union of multiple bbox geoms")
+            geometry = unary_union(geometry)
+
+        if geometry.is_empty:
+            msg = f"No bbox extracted from {geometry}"
+            log.error(msg)
+            raise ValueError(msg) from None
+
+        bbox = geometry.bounds
+        # left, bottom, right, top
+        # minX, minY, maxX, maxY
+        return bbox
+
+
+class StringBoundaryHandler(BoundaryHandler):
+    """Extracts BBOX from string representation."""
+
+    def __init__(self, boundary: str):
+        """Initialize the StringBoundaryHandler with a BoundaryHandler input."""
+        self.boundary = boundary
+
+    def make_bbox(self) -> BoundingBox:
+        """A function to parse BBOX string."""
+        try:
+            if "," in self.boundary:
+                bbox_parts = self.boundary.split(",")
+            else:
+                bbox_parts = self.boundary.split(" ")
+            bbox = tuple(float(x) for x in bbox_parts)
+            if len(bbox) == 4:
+                # BBOX valid
+                return bbox
+            else:
+                msg = f"BBOX string malformed: {bbox}"
+                log.error(msg)
+                raise ValueError(msg) from None
+        except Exception as e:
+            log.error(e)
+            msg = f"Failed to parse BBOX string: {self.boundary}"
+            log.error(msg)
+            raise ValueError(msg) from None
 
 
 def dlthread(
@@ -272,6 +378,7 @@ class BaseMapper(object):
         else:
             log.debug("%s doesn't exists" % filespec)
             return False
+
 
 def tileid_from_xyz_dir_path(filepath: Union[Path, str], is_xy: bool = False) -> int:
     """Helper function to get the tile id from a tile in xyz directory structure.
